test(extract): test output of a cruise wich reachability against the json schema
