+++ test/extract/fixtures/reachable/src/index.js
@@ -1,0 +1,3 @@
+const things = require('./relevanty');
+
+console.log(things);
