+++ test/extract/fixtures/reachable/src/relevanty/index.js
@@ -1,0 +1,1 @@
+module.exports = 3;
