+++ test/extract/fixtures/reachable/src/untouched-cluster/index.js
@@ -1,0 +1,3 @@
+const hoodoo = require('./untouched-one');
+
+module.exports = 2* hoodoo;
