+++ test/extract/fixtures/reachable/src/untouched-cluster/untouched-one.js
@@ -1,0 +1,1 @@
+module.exports = 2 * require('./untouched-two')
