+++ test/extract/fixtures/reachable/src/untouched-cluster/untouched-two.js
@@ -1,0 +1,1 @@
+module.exports = 2;
