+++ test/extract/index.maxdepth.spec.js
@@ -1,0 +1,38 @@
+const chai                    = require('chai');
+const extract                 = require('../../src/extract');
+const depSchema               = require('../../src/extract/results-schema.json');
+const normalize               = require('../../src/main/options/normalize');
+const normalizeResolveOptions = require('../../src/main/resolveOptions/normalize');
+
+const expect = chai.expect;
+
+chai.use(require('chai-json-schema'));
+
+describe('extract/index - Max depth', () => {
+    /* eslint no-magic-numbers:0 */
+    [0, 1, 2, 4].forEach((pDepth) =>
+        it(`returns the correct graph when max-depth === ${pDepth}`, () => {
+            const lOptions = normalize({
+                maxDepth: pDepth
+            });
+            const lResolveOptions = normalizeResolveOptions(
+                {
+                    bustTheCache: true
+                },
+                lOptions
+            );
+            const lResult = extract(
+                ["./test/extract/fixtures/maxDepth/index.js"],
+                lOptions,
+                lResolveOptions
+            );
+            /* eslint import/no-dynamic-require:0, security/detect-non-literal-require:0 */
+
+            expect(lResult.modules).to.deep.equal(
+                require(`./fixtures/maxDepth${pDepth}.json`).modules
+            );
+            expect(lResult).to.be.jsonSchema(depSchema);
+        })
+    );
+});
+/* eslint global-require: 0*/
