+++ test/extract/index.spec.js
@@ -1,4 +1,3 @@
-const fs                      = require('fs');
 const chai                    = require('chai');
 const extract                 = require('../../src/extract');
 const depSchema               = require('../../src/extract/results-schema.json');
@@ -50,156 +49,4 @@ describe(
     () => coffeeRecursiveFixtures.forEach(runRecursiveFixture)
 );
 
-describe('extract/index - Max depth', () => {
-    /* eslint no-magic-numbers:0 */
-    [0, 1, 2, 4].forEach((pDepth) =>
-        it(`returns the correct graph when max-depth === ${pDepth}`, () => {
-            const lOptions = normalize({
-                maxDepth: pDepth
-            });
-            const lResolveOptions = normalizeResolveOptions(
-                {
-                    bustTheCache: true
-                },
-                lOptions
-            );
-            const lResult = extract(
-                ["./test/extract/fixtures/maxDepth/index.js"],
-                lOptions,
-                lResolveOptions
-            );
-            /* eslint import/no-dynamic-require:0, security/detect-non-literal-require:0 */
-
-            expect(lResult.modules).to.deep.equal(
-                require(`./fixtures/maxDepth${pDepth}.json`).modules
-            );
-            expect(lResult).to.be.jsonSchema(depSchema);
-        })
-    );
-});
-
-describe('extract/index - do not follow', () => {
-
-    it('do not follow - doNotFollow.path', () => {
-        const lOptions = normalize({
-            doNotFollow: {
-                path: "donotfollowonceinthisfolder"
-            }
-        });
-        const lResolveOptions = normalizeResolveOptions(
-            {
-                bustTheCache: true
-            },
-            lOptions
-        );
-        const lResult = extract(
-            ["./test/extract/fixtures/donotfollow/index.js"],
-            lOptions,
-            lResolveOptions
-        );
-
-        expect(lResult.modules).to.deep.equal(
-            require('./fixtures/donotfollow.json').modules
-        );
-        expect(lResult).to.be.jsonSchema(depSchema);
-    });
-
-    it('do not follow - doNotFollow.dependencyTypes', () => {
-        const lOptions = normalize({
-            doNotFollow: {
-                dependencyTypes: ["npm-no-pkg"]
-            }
-        });
-        const lResolveOptions = normalizeResolveOptions(
-            {
-                bustTheCache: true
-            },
-            lOptions
-        );
-        const lResult = extract(
-            ["./test/extract/fixtures/donotfollow-dependency-types/index.js"],
-            lOptions,
-            lResolveOptions
-        );
-
-        expect(lResult.modules).to.deep.equal(
-            require('./fixtures/donotfollow-dependency-types.json').modules
-        );
-        expect(lResult).to.be.jsonSchema(depSchema);
-    });
-});
-
-
-describe('extract/index - cache busting', () => {
-    it('delivers a different output', () => {
-        const lOptions = normalize({
-            ruleSet: {
-                forbidden: [
-                    {
-                        name: "burp-on-core",
-                        severity: "error",
-                        from: {},
-                        to: {
-                            dependencyTypes: ["core"]
-                        }
-                    }
-                ]
-            },
-            validate: true,
-            tsPreCompilationDeps: true,
-            doNotFollow: {
-                dependencyTypes: [
-                    "npm",
-                    "npm-dev",
-                    "npm-optional",
-                    "npm-peer",
-                    "npm-bundled"
-                ]
-            }
-
-        });
-        const lResolveOptions = normalizeResolveOptions(
-            {},
-            lOptions
-        );
-        const lFirstResultFixture = require('./fixtures/cache-busting-first-tree.json');
-        const lSecondResultFixture = require('./fixtures/cache-busting-second-tree.json');
-
-        fs.renameSync(
-            "./test/extract/fixtures/cache-busting-first-tree",
-            "./test/extract/fixtures/cache-busting"
-        );
-
-        const lFirstResult = extract(
-            ["./test/extract/fixtures/cache-busting/index.ts"],
-            lOptions,
-            lResolveOptions
-        );
-
-        fs.renameSync(
-            "./test/extract/fixtures/cache-busting",
-            "./test/extract/fixtures/cache-busting-first-tree"
-        );
-
-        fs.renameSync(
-            "./test/extract/fixtures/cache-busting-second-tree",
-            "./test/extract/fixtures/cache-busting"
-        );
-
-        const lSecondResult = extract(
-            ["./test/extract/fixtures/cache-busting/index.ts"],
-            lOptions,
-            lResolveOptions
-        );
-
-        fs.renameSync(
-            "./test/extract/fixtures/cache-busting",
-            "./test/extract/fixtures/cache-busting-second-tree"
-        );
-
-        expect(lFirstResult).to.deep.equal(lFirstResultFixture);
-        expect(lSecondResult).to.deep.equal(lSecondResultFixture);
-        expect(lSecondResult).to.not.deep.equal(lFirstResult);
-    });
-});
 /* eslint global-require: 0*/
